Update: Disable api type arg for now
